// src/types/asset/Asset.ts
export type Asset = {
  stop_id: number;
  route_number: number;
  city: string;
  name: string;
  latitude: number;
  longitude: number;
  compliant: string;
  shelter_rating: string;
  shelter_type: string;
  bench_rating: string;
  bench_type: string;
  trash_can_type: string;
  trash_can_rating: string;
};